% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_tempo.R
\name{avg_tempo}
\alias{avg_tempo}
\title{Average Tempo}
\usage{
avg_tempo(df, tempo, performer, year, add_flex = F)
}
\arguments{
\item{df}{The dataframe}

\item{tempo}{Whatever your tempo column is called.}

\item{year}{Whatever your year column is called.}

\item{grouping}{grouping operator}

\item{selection}{Filtered selection.}
}
\description{
This function provides an easy eay to get the average tempo by a number of groups.
You can look at tempo by measure, beat, composer, or piece. It returns a dataframe, not a tibble.
}
\author{
Daniel Shanahan
}
