% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{import_all}
\alias{import_all}
\title{Import all tempo files from a folder.}
\usage{
import_all(filetype)
}
\arguments{
\item{filetype}{The type of files in your directory. Currently need to be tab delimited.
Imports all tempo files from a certain folder.
Converts tempo data if needed.}
}
\description{
Import all tempo files from a folder.
}
