% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazurka_op17_no4.R
\docType{data}
\name{mazurka_op17_no4}
\alias{mazurka_op17_no4}
\title{Tap tempo data for Chopin's Mazurka, op.17, no.4.}
\format{
A dataset that includes 24,453 rows and 5 variables.
}
\source{
\url{http://www.mazurka.org.uk/info/excel/beat/}
}
\usage{
mazurka_op17_no4
}
\description{
Tap tempo data for Chopin's Mazurka, op.17, no.4.
}
\keyword{datasets}
