% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_select.R
\name{measure_select}
\alias{measure_select}
\title{Select Measures}
\usage{
measure_select(df, from, to, measure.beat = "measure.beat")
}
\arguments{
\item{df}{The dataframe}

\item{from}{where you want the selection to begin}

\item{to}{where you want the selection to end}

\item{measure.beat}{what your measure column is called.}
}
\description{
This is basically a wrapping of the dplyr filter program, but a little more straightforward (and therefore less flexible and powerful).
}
\author{
Daniel Shanahan
}
