% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2bpm.R
\name{ms2bpm}
\alias{ms2bpm}
\title{Milliseconds to Beats per Minute.}
\usage{
ms2bpm(df, ms, cumulative = F, seconds = T)
}
\arguments{
\item{ms}{millisecond column name. This can be set up to either be consistent onsets between beats,
or a cumulative running total of milliseconds.  Which one it is should determine how you use the cumulative argument.}

\item{cumulative}{boolean that goes between a cumulative onset count,
as Sonic Visualizer will output, and one that is already
using the time between onsets in a non-cumulative way.}

\item{data}{dataframe being examined}
}
\description{
This function converts the millisecond data into beats per minute data.
}
\author{
Daniel Shanahan
}
