% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_flex.R
\name{tempo_flex}
\alias{tempo_flex}
\title{This function provides a flexibility rating (as measured with a standard
deviation of the entire unit given). It uses dplyr, and requires the year.}
\usage{
tempo_flex(data, performer = "performer", tempo = "tempo")
}
\arguments{
\item{tempo}{The bpm data; it's own column.}

\item{flex}{The flexibility rating; a standard deviation of the tempi.}

\item{year}{The year column.}
}
\description{
This function provides a flexibility rating (as measured with a standard
deviation of the entire unit given). It uses dplyr, and requires the year.
}
