% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{tempo_function}
\alias{tempo_function}
\alias{test_function}
\title{test function}
\usage{
test_function(name)
}
\arguments{
\item{name, }{who to say hello to.}
}
\description{
Just seeing if this can be downloaded by students.
}
