% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo_over_time.R
\name{tempo_over_time}
\alias{tempo_over_time}
\title{Plots average tempo and flexibility over time.}
\usage{
tempo_over_time(
  data,
  method = "lm",
  big_title = "Data over Time",
  brief_subtitle = "Average Tempo and Flexibility",
 
    small_caption = "Flexibility is measured as the standard\\n                            deviation of the entire unit."
)
}
\arguments{
\item{data, }{The dataframe}

\item{year, }{The year, or any other timeframe variable. Required.}

\item{average_tempo, }{The average tempo, as provided in the dataframe.}

\item{flex, }{flexibility, as determined by the standard deviation of the entire unit.}
}
\description{
This function just uses one of my most commonly
used ggplot graphs, plotting tempo over time.
}
